refactor(patina): split oversized files to satisfy source-length gate
